-Вынес модульные тесты в отдельный файл -Мелкие стилистические исправления
